reorganized the code to be a little easier to use
