+++ package.json
@@ -29,7 +29,6 @@
     "boom": "^7.2.0",
     "cfonts": "^2.1.3",
     "commander": "^2.16.0",
-    "figlet": "^1.2.0",
     "hawk": "^7.0.7",
     "hoek": "^5.0.3",
     "inquirer": "^6.0.0",
